requirements.txt
alabaster==0.7.12
asgiref==3.2.7
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
Babel==2.8.0
bandit==1.6.2
bleach==3.1.5
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
codecov==2.1.6
colorama==0.4.3
coverage==5.1
Django==3.0.7
django-filter==2.3.0
docutils==0.16
dparse==0.5.1
gitdb==4.0.5
GitPython==3.1.3
idna==2.9
imagesize==1.2.0
isort==4.3.21
Jinja2==2.11.2
keyring==21.2.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.3.0
mypy==0.780
mypy-extensions==0.4.3
packaging==20.4
pbr==5.4.5
pkginfo==1.5.0.1
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pycodestyle==2.6.0
pydocstyle==5.0.2
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
pytz==2020.1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
readme-renderer==26.0
requests==2.23.0
requests-toolbelt==0.9.1
safety==1.9.0
six==1.15.0
smmap==3.0.4
snowballstemmer==2.0.0
Sphinx==3.1.1
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
stevedore==2.0.0
toml==0.10.1
tqdm==4.46.1
twine==3.1.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
wcwidth==0.2.4
webencodings==0.5.1
wrapt==1.12.1
